fix build warnings

+++ Classes/Views/PBGitRevisionCell.m
@@ -113,7 +113,7 @@ const BOOL SHUFFLE_COLORS = NO;
 
 - (void) drawCircleInRect: (NSRect) r
 {
-	int c = cellInfo.position;
+	int c = (int)cellInfo.position;
 	NSPoint origin = r.origin;
 	NSPoint columnOrigin = { origin.x + COLUMN_WIDTH * c, origin.y};
 
@@ -149,7 +149,7 @@ const BOOL SHUFFLE_COLORS = NO;
 
 - (void) drawTriangleInRect: (NSRect) r sign: (char) sign
 {
-	int c = cellInfo.position;
+	int c = (int)cellInfo.position;
 	int columnHeight = 10;
 	int columnWidth = 8;
